split onboarding controllers into multiple files
