+++ CareForMe/Features/Connect Users/Controller/OnboardingPagedViewController.swift
@@ -1,73 +1,12 @@
 //
-//  ParchmentPageViewController.swift
+//  OnboardingPagedViewController.swift
 //  CareForMe
 //
-//  Created by Kenneth Dubroff on 8/10/21.
+//  Created by Kenneth Dubroff on 8/11/21.
 //
 
 import UIKit
 import Parchment
-
-struct ParchmentCellTest: PagingItem, Hashable, Comparable {
-    
-    var identifier: Int
-    
-    static func < (lhs: ParchmentCellTest, rhs: ParchmentCellTest) -> Bool {
-        lhs.identifier < rhs.identifier
-    }
-    
-}
-
-class OnboardingViewController: InstructionViewController {
-    
-    let indicatorTitle: String
-    let id: Int
-    
-    init(id: Int, indicatorText: String, title: String, instructions: String, image: Gif?, caption: String? = nil, buttonTitle: String, selectionDelegate: TargetSelector, additionalViews: [UIView]) {
-        
-        self.indicatorTitle = indicatorText
-        self.id = id
-        super.init(title: title, instructions: instructions, image: image, buttonTitle: buttonTitle, selectionDelegate: selectionDelegate)
-        instructionView.button.tag = id
-
-        for view in additionalViews {
-            instructionView.addView(view)
-        }
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("programmatic ViewController")
-    }
-}
-
-final class OnboardingLinkVC: OnboardingViewController {
-    lazy var linkButton: UIButton = .fullWidthButton(with: "I want to confirm a code instead", color: .named(.secondaryLink), targetAndSelector: TargetSelector(target: self, selector: #selector(displayCodeConfirmationViewController)))
-    
-    lazy var codeLabel: UILabel = {
-        let codeLabel = UILabel()
-        codeLabel.text = Int.randomString()
-        codeLabel.font = .boldSystemFont(ofSize: 36)
-        codeLabel.textAlignment = .center
-        return codeLabel
-    }()
-    
-    @objc private func displayCodeConfirmationViewController() {
-        let vc = ConfirmPINViewController()
-        showDetailViewController(vc, sender: nil)
-    }
-    
-    init(selectionDelegate: TargetSelector, additionalViews: [UIView] = []) {
-        
-        super.init(id: 2, indicatorText: "Link", title: "Link To Your Companion", instructions: "Linking to a companion is easy. Just provide them with this 6 digit code and ask them to download the app", image: nil, buttonTitle: "Let's Get Started!", selectionDelegate: selectionDelegate, additionalViews: additionalViews)
-        
-        instructionView.addView(codeLabel)
-        instructionView.addView(linkButton)
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("programmatic view controller")
-    }
-}
 
 class OnboardingPagedViewController: PagingViewController {
